Renamed package_lists to constants as we can use that for more than just package lists in the future
